Use dein.vim
The dein.vim file in this commit is a symlink to ~/src/dein.vim
so that I don't have to include that repo in here as a submodule.

I would do it with subrepo, but dein needs to have the ability to manage
the dein plugin itself, and it does not currently support subrepo.